+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add a run operator `A.expect(P)` to evaluate the result of `A` with `P` (#1303)
+- Add a run operator `A.expect(P)` to test the state predicate `P` in the state resulting from applying action `A` (#1303)
 
 ### Changed
 ### Deprecated