+++ CHANGELOG.md
@@ -8,6 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## UNRELEASED
 
 ### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+## v0.18.0 -- 2024-01-03
+
+### Added
 
 - Add a run operator `A.expect(P)` to test the state predicate `P` in the state resulting from applying action `A` (#1303)